backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
pytrends
openai-whisper
google-genai
requests
pandas
python-multipart
aiofiles
